# build/swagger.json
{
    "components": {
...
        "schemas": {
...
            "UserDto": {
                "properties": {
                    "id": {
                        "type": "number",
                        "format": "double"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "activated": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "firstName",
                    "lastName",
                    "email",
                    "phoneNumber",
                    "activated"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "Token": {
                "properties": {
                    "accessToken": {
                        "type": "string"
                    }
                },
                "required": [
                    "accessToken"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "UpdateUserDto": {
                "properties": {
                    "login": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    }
                },
                "required": [
                    "login",
                    "password",
                    "firstName",
                    "lastName",
                    "email",
                    "phoneNumber"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "LoginDto": {
                "properties": {
                    "login": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "login",
                    "password"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "UserRole": {
                "enum": [
                    "admin",
                    "worker",
                    "client"
                ],
                "type": "string"
            },
            "CreateUserDto": {
                "properties": {
                    "login": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "role": {
                        "$ref": "#/components/schemas/UserRole"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    }
                },
                "required": [
                    "login",
                    "password",
                    "role",
                    "firstName",
                    "lastName",
                    "email",
                    "phoneNumber"
                ],
                "type": "object",
                "additionalProperties": false
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "info": {
        "title": "car_repair_station",
        "version": "1.0.0",
        "description": "Back-END car repair station",
        "license": {
            "name": "ISC"
        },
        "contact": {
            "name": "Stawr28@gmail.com"
        }
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/users": {
            "get": {
                "operationId": "GetAllUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/UserDto"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "admin",
                            "worker"
                        ]
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "CreateUser",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserDto"
                            }
                        }
                    }
                }
            }
        },
        "/users/{userId}": {
            "get": {
                "operationId": "GetUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "admin",
                            "worker"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            },
            "delete": {
                "operationId": "DeleteconcreteUser",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "admin",
                            "worker"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateConcreteUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "admin",
                            "worker"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        }
                    }
                }
            }
        },
        "/users/get/me": {
            "get": {
                "operationId": "GetMyUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "client"
                        ]
                    }
                ],
                "parameters": []
            },
            "put": {
                "operationId": "UpdateMyUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "client"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteMyUser",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "client"
                        ]
                    }
                ],
                "parameters": []
            }
        },
        "/users/login": {
            "post": {
                "operationId": "Login",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDto"
                            }
                        }
                    }
                }
            }
        },
        "/users/signup-by-link/{link}": {
            "get": {
                "operationId": "SignUpByLink",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "link",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/users/logout": {
            "post": {
                "operationId": "LogOut",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "admin",
                            "worker",
                            "client"
                        ]
                    }
                ],
                "parameters": []
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}